requirements.txt
streamlit
google-generativeai
PyPDF2
langchain-core
langchain-ollama
